Split off many commands into plugin files
Organized by type of command for the most part
