+++ bot_command.py
@@ -3,7 +3,6 @@ import asyncio
 import command_list as cmdlst
 import sys
 import importlib
-import mtg_card_cmd
 import os
 
 from enum import Enum
@@ -101,24 +100,11 @@ def init_commands(client):
     cmds = []
 
     cmds.append(bot_cmd("!reload", None, cmd_lvl.bot_admins, 'Reload the functionality of the commands'))
-    cmds.append(bot_cmd("!help", cmdlst.help, 1, 'Show a user all the commands they can use'))
-    cmds.append(bot_cmd("!ping", cmdlst.ping, 1, 'If I\'m alive I will say Pong!'))
-    cmds.append(bot_cmd("!noise", cmdlst.noise, 1, 'Generate 20 random letters'))
-    cmds.append(bot_cmd("!deleteme", cmdlst.deleteme, cmd_lvl.mods, 'Delete all messages from the bot, use -s to not get feedback'))
-    cmds.append(bot_cmd("!deletecmds", cmdlst.deletecmds, 1, 'Delete all command messages written by the user'))
-    cmds.append(bot_cmd("!deleteallcmds", cmdlst.deleteallcmds, cmd_lvl.mods, 'Delete all command messages written by anyone'))
-    cmds.append(bot_cmd("!join", cmdlst.join, cmd_lvl.admins, 'Join the voice channel the user is in'))
-    cmds.append(bot_cmd("!leave", cmdlst.leave, cmd_lvl.admins, 'Leave whatever voice channel the bot is in'))
     cmds.append(bot_cmd("!killme", cmdlst.endbot, cmd_lvl.bot_admins, 'Turn off the bot, will need to be manually restarted'))
-    cmds.append(bot_cmd("!downloadchat", cmdlst.downloadhistory, cmd_lvl.bot_admins, 'Downloads chat history since the last download in the current channel'))
     cmds.append(bot_cmd("!rock", cmdlst.rockfact, 1, 'Say a random rock fact!'))
-    cmds.append(bot_cmd("!flip", cmdlst.coinflip, 1, 'Flip a coin with cryptographically secure randomness!'))
-    cmds.append(bot_cmd("!roulette", cmdlst.roulette, 1, 'Test your luck!'))
-    cmds.append(bot_cmd("!markov", cmdlst.markov, 1, 'Generate a random sentence based on the current markov data (set by admins)'))
-    cmds.append(bot_cmd("!markovusers", cmdlst.markovusers, 1, 'Get sent a list of markov users'))
+    
     client.markov_chains = None
     client.named_markov_chains = None
-    cmds.append(bot_cmd("!cardboard", mtg_card_cmd.mtgcard, 1, 'Search for a piece of cardboard by name'))
 
     modules = load_plugins()
     for m in modules.keys():
